requirements.txt
fastapi>=0.111,<1.0
uvicorn>=0.30,<1.0
sqlalchemy>=2.0,<3.0
psycopg[binary]>=3.1,<4.0
pydantic-settings>=2.0,<3.0

pytest>=8.0,<9.0
httpx>=0.27,<1.0